requirements.txt
annotated-types==0.6.0
anyio==4.2.0
blinker==1.6.3
click==8.1.7
distlib==0.3.7
django-environ==0.11.2
fastapi==0.108.0
filelock==3.13.1
Flask==3.0.0
h11==0.14.0
idna==3.6
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
mysql-connector==2.2.9
platformdirs==4.0.0
psycopg2==2.9.9
pydantic==2.5.3
pydantic_core==2.14.6
sniffio==1.3.0
SQLAlchemy==2.0.25
starlette==0.32.0.post1
typing_extensions==4.9.0
uvicorn==0.25.0
virtualenv==20.24.7
virtualenvwrapper-win==1.2.7
Werkzeug==3.0.0
